Added next steps for future improvement

+++ src/main/java/com/example/demo/services/BookService.java
@@ -12,21 +12,34 @@ import java.util.List;
 @Service
 public class BookService {
 
-    @Autowired
-    private BookRepository bookRepository;
 
+    private final BookRepository bookRepository;
+    private final ElasticSearchService elasticSearchService;
     @Autowired
     private BookMapper bookMapper;
+
+    public BookService(BookRepository bookRepository, ElasticSearchService elasticSearchService) {
+        this.bookRepository = bookRepository;
+        this.elasticSearchService = elasticSearchService;
+    }
+
+    public Book saveBook(Book book) {
+        Book savedBook = bookRepository.save(book);
+        elasticSearchService.indexBook(savedBook);
+        return savedBook;
+    }
 
     public List<Book> getAllBooks() {
         return (List<Book>) bookRepository.findAll();
     }
 
-    public BookDTO createBook(BookDTO reviewDTO) {
-        Book review = bookMapper.toEntity(reviewDTO);
-        Book savedBook = bookRepository.save(review);
+    public BookDTO createBook(BookDTO bookDTO) {
+        Book book = bookMapper.toEntity(bookDTO);
+        Book savedBook = bookRepository.save(book);
+        elasticSearchService.indexBook(savedBook);
         return bookMapper.toDto(savedBook);
     }
+
 
     public BookDTO getBookById(Long id) {
         Book review = bookRepository.findById(id)
@@ -34,23 +47,26 @@ public class BookService {
         return bookMapper.toDto(review);
     }
 
+
     public BookDTO updateBook(Long id, BookDTO updatedBookDTO) {
         Book book = bookRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Book not found"));
 
-        Book bookToUpdate = BookMapper.INSTANCE.toEntity(updatedBookDTO);
+        bookMapper.updateBookFromDto(updatedBookDTO, book);
 
-        book.setAuthor(bookToUpdate.getAuthor());
-        book.setAvailable(bookToUpdate.getAvailable());
-        book.setCategories(bookToUpdate.getCategories());
-        book.setDescription(bookToUpdate.getDescription());
-        book.setLanguage(bookToUpdate.getLanguage());
-        book.setTitle(bookToUpdate.getTitle());
+        Book updatedBook = bookRepository.save(book);
 
-        return BookMapper.INSTANCE.toDto(bookRepository.save(book));
+        elasticSearchService.indexBook(updatedBook);
+
+        return bookMapper.toDto(updatedBook);
     }
 
     public void deleteBook(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new RuntimeException("Book not found");
+        }
         bookRepository.deleteById(id);
+        elasticSearchService.deleteBook(id);
     }
+
 }
